+++ src/components/badgedLink/index.js
@@ -1,0 +1,3 @@
+import BadgedLink from './BadgedLink';
+
+export default BadgedLink;