+++ src/components/link/link.stories.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { addStoryInGroup, LOW_LEVEL_BLOCKS } from '../../../.storybook/utils';
-import { IconAddSmallOutline } from '@teamleader/ui-icons';
 import Link from './Link';
 
 export default {
@@ -8,11 +7,7 @@ export default {
   title: addStoryInGroup(LOW_LEVEL_BLOCKS, 'Link'),
 };
 
-export const DefaultStory = (args) => (
-  <Link {...args} icon={<IconAddSmallOutline />}>
-    I am a link with an icon
-  </Link>
-);
+export const DefaultStory = (args) => <Link {...args}>I am a link</Link>;
 
 DefaultStory.args = {
   href: 'https://www.teamleader.be',