+++ src/components/utils/utils.js
@@ -6,11 +6,3 @@ export const getViewport = () => ({
 export const isElementOverflowingX = (element) => element.clientWidth < element.scrollWidth;
 
 export const isElementOverflowingY = (element) => element.clientHeight < element.scrollHeight;
-
-export const elementIsDark = (color, dark) => {
-  const lightColors = ['white', 'neutral'];
-  if (lightColors.includes(color)) {
-    return dark;
-  }
-  return false;
-};