+++ CHANGELOG.md
@@ -12,6 +12,29 @@
 
 ### Dependency updates
 
+## [12.0.0] - 2022-01-25
+
+### Added
+
+- `BadgedLink`: ([@qubis741](https://github.com/qubis741) in [#1925](https://github.com/teamleadercrm/ui/pull/1925))
+
+### Removed
+
+- [BREAKING] `Pagination, Button`: Removed the `inverse` property ([@ArnaudWeyts](https://github.com/ArnaudWeyts) in [#1890](https://github.com/teamleadercrm/ui/pull/1890))
+- [BREAKING] `StatusBullet, TimerPulser, DatePickerRange, DatePickerInputRange`: ([@qubis741](https://github.com/qubis741) in [#1889](https://github.com/teamleadercrm/ui/pull/1889))
+- [BREAKING] `ScrollContainer`: ([@qubis741](https://github.com/qubis741) in [#1891](https://github.com/teamleadercrm/ui/pull/1891))
+- [BREAKING] `LoadingMolecule`: ([@qubis741](https://github.com/qubis741) in [#1888](https://github.com/teamleadercrm/ui/pull/1888))
+- `IslandGroup`, `Island`, `Section`: removed `dark` property ([@ArnaudWeyts](https://github.com/ArnaudWeyts) in [#1892](https://github.com/teamleadercrm/ui/pull/1892))
+- [BREAKING] `Link`: removed `icon` and `badged` property ([@qubis741](https://github.com/qubis741) in [#1925](https://github.com/teamleadercrm/ui/pull/1925))
+
+### Fixed
+
+- `Button`: fixed `Spinner` size when `Button` size is `tiny` ([@qubis741](https://github.com/qubis741) in [#1931](https://github.com/teamleadercrm/ui/pull/1931))
+
+### Dependency updates
+
+- Upgraded all dependencies to latest except `react`, `react-dom`, `react-day-picker` (latest is beta), `postcss-preset-env`, `postcss-each` ([@qubis741](https://github.com/qubis741) in [#1893](https://github.com/teamleadercrm/ui/pull/1893))
+
 ## [11.0.0] - 2021-12-20
 
 ### Added
@@ -1985,7 +2008,6 @@
 ### Changed
 
 - `Label`: changed the `css display rule` to use the `display prop` of our Box component.
-
 - `Dependencies`:
   - `eslint-plugin-node` from `^6.0.0` to `^9.1.0`
   - `eslint-config-standard` from `^11.0.0` to `^13.0.1`